% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_plotting_func.R
\name{cal_phi_0}
\alias{cal_phi_0}
\title{Calculate \phi_0 from AR model.}
\usage{
cal_phi_0(ar_mod, ord, digits = 6)
}
\arguments{
\item{ar_mod}{An AR model.}

\item{ord}{The order of AR model.}

\item{digits}{The number of digits to show float number.}
}
\description{
Calculate \phi_0 from AR model.
}
