% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_plotting_func.R
\name{draw_arima_forecast_fig}
\alias{draw_arima_forecast_fig}
\title{Draw an ARIMA model and plot forecast results.}
\usage{
draw_arima_forecast_fig(
  da_ts,
  eotr,
  h,
  npts,
  frequency,
  ts_fc_res,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL
)
}
\arguments{
\item{da_ts}{A ts object.}

\item{eotr}{The end index of training data.}

\item{h}{The forecast horizon.}

\item{npts}{The number of points at the end of training data to plot.}

\item{frequency}{The frequency of the ts object.}

\item{ts_fc_res}{An object already fit.}

\item{main}{The title of the figure.}

\item{xlab}{xlab.}

\item{ylab}{ylab.}

\item{ylim}{ylim.}
}
\description{
Draw an ARIMA model and plot forecast results.
}
