% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_plotting_func.R
\name{get_mu}
\alias{get_mu}
\title{Get \mu from ARIMA model.}
\usage{
get_mu(arima_mod, digits = 6)
}
\arguments{
\item{arima_mod}{An ARIMA model.}

\item{digits}{The number of digits to show float number.}
}
\description{
Get \mu from ARIMA model.
}
