% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_plotting_func.R
\name{install_redir_output}
\alias{install_redir_output}
\title{Install a package and direct output to a log file.}
\usage{
install_redir_output(lib, log_fpath)
}
\arguments{
\item{lib}{A string of library name.}

\item{log_fpath}{A log file path to save installation output.}
}
\description{
Install package and direct output to a log file.
}
\examples{
install_redir_output("TSA", TSA_output.log)
}
