% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_plotting_func.R
\name{perform_and_print_eacf}
\alias{perform_and_print_eacf}
\title{Perform and print eacf of a time-series.}
\usage{
perform_and_print_eacf(da, ar.max, ma.max)
}
\arguments{
\item{da}{A data series.}

\item{ar.max}{An eacf param.}

\item{ma.max}{An eacf param.}
}
\description{
Perform and print eacf of a time-series.
}
