% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_plotting_func.R
\name{plot_acf}
\alias{plot_acf}
\title{Plot acf of a time-series.}
\usage{
plot_acf(da, lag.max = NULL, main = NULL, w = NULL, h = NULL, ...)
}
\arguments{
\item{da}{A data series.}

\item{lag.max}{An acf param.}

\item{main}{The title of the figure.}

\item{w}{Plot width.}

\item{h}{Plot height.}
}
\description{
Plot acf of a time-series.
}
