% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_plotting_func.R
\name{plot_auto_arima_forecast_fig}
\alias{plot_auto_arima_forecast_fig}
\title{Get \mu from ARIMA model.}
\usage{
plot_auto_arima_forecast_fig(
  da_ts,
  eotr,
  h,
  npts,
  frequency,
  xreg = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{da_ts}{A ts object.}

\item{eotr}{The end index of training data.}

\item{h}{The forecast horizon.}

\item{npts}{The number of points at the end of training data to plot.}

\item{frequency}{The frequency of the ts object.}

\item{xreg}{arima param (data.frame).}

\item{main}{The title of the figure.}

\item{xlab}{xlab.}

\item{ylab}{ylab.}

\item{...}{auto.arima params.}
}
\description{
Get \mu from ARIMA model.
}
