% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_plotting_func.R
\name{plot_forecast_fig}
\alias{plot_forecast_fig}
\title{Get \mu from ARIMA model.}
\usage{
plot_forecast_fig(
  da_ts,
  eotr,
  h,
  npts,
  frequency,
  order,
  fixed,
  method,
  transform_pars,
  main,
  xlab,
  ylab,
  ylim = NULL
)
}
\arguments{
\item{da_ts}{A ts object.}

\item{eotr}{The end index of training data.}

\item{h}{The forecast horizon.}

\item{npts}{The number of points at the end of training data to plot.}

\item{frequency}{The frequency of the ts object.}

\item{order}{ARIMA param.}

\item{fixed}{ARIMA param.}

\item{method}{ARIMA param.}

\item{transform_pars}{ARIMA param.}

\item{main}{The title of the figure.}

\item{xlab}{xlab.}

\item{ylab}{ylab.}
}
\description{
Get \mu from ARIMA model.
}
