% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_plotting_func.R
\name{plot_pacf_acf}
\alias{plot_pacf_acf}
\title{Plot pacf and acf of a time-series.}
\usage{
plot_pacf_acf(da, lag.max = NULL, main = NULL, w = NULL, h = NULL, ...)
}
\arguments{
\item{da}{An data series.}

\item{lag.max}{An acf param.}

\item{main}{The title of the figure.}

\item{w}{Plot width.}

\item{h}{Plot height.}
}
\description{
Plot pacf and acf of a time-series.
}
