% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_plotting_func.R
\name{plot_time_fig}
\alias{plot_time_fig}
\title{Plot time figure}
\usage{
plot_time_fig(ts, main = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{ts}{A ts object.}

\item{main}{The title of the figure.}

\item{xlab}{xlab.}

\item{ylab}{ylab.}
}
\description{
Plot time figure
}
