% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_plotting_func.R
\name{tsdiag1}
\alias{tsdiag1}
\title{Perform diagnosis on models (my version).}
\usage{
tsdiag1(object, gof.lag, ...)
}
\arguments{
\item{object}{Model.}

\item{gof.lag}{The lag.max.}

\item{...}{acf params.}
}
\description{
Perform diagnosis on models (my version).
}
\examples{
tsdiag1(ew_lrtn_m1, gof.lag = 36, `drop.lag.0` = FALSE)
tsdiag1(ew_lrtn_m1, gof.lag = 36, `drop.lag.0` = TRUE) 
}
