% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_plotting_func.R
\name{perform_and_print_eacf}
\alias{perform_and_print_eacf}
\title{Perform and print eacf of a time-series.}
\usage{
perform_and_print_eacf(da, ar.max, ma.max)
}
\arguments{
\item{da}{An data series.}

\item{lag.max}{An acf param.}

\item{main}{The title of the figure.}

\item{w}{Plot width.}

\item{h}{Plot height.}
}
\description{
Perform and print eacf of a time-series.
}
