% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_plotting_func.R
\name{plot_arima_forecast_fig}
\alias{plot_arima_forecast_fig}
\title{Fit a ARIMA model and plot forecast results.}
\usage{
plot_arima_forecast_fig(
  da_ts,
  eotr,
  h,
  npts,
  frequency,
  order,
  seasonal,
  fixed,
  method,
  include.mean,
  transform.pars,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  ts_fc_res = NULL
)
}
\arguments{
\item{da_ts}{A ts object.}

\item{eotr}{The end index of training data.}

\item{h}{The forecast horizon.}

\item{npts}{The number of points at the end of training data to plot.}

\item{frequency}{The frequency of the ts object.}

\item{order}{arima param.}

\item{seasonal}{arima param.}

\item{fixed}{arima param.}

\item{method}{arima param.}

\item{include.mean}{arima param.}

\item{transform.pars}{arima param.}

\item{main}{The title of the figure.}

\item{xlab}{xlab.}

\item{ylab}{ylab.}

\item{ylim}{ylim.}

\item{ts_fc_res}{An object already fit.}
}
\description{
Fit a ARIMA model and plot forecast results.
}
