% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_plotting_func.R
\name{plot_pacf_acf}
\alias{plot_pacf_acf}
\title{Plot pacf and acf of a time-series.}
\usage{
plot_pacf_acf(da, freq = NA, ...)
}
\arguments{
\item{da}{A data series.}

\item{freq}{A frequency in taking seasonal difference.}
}
\description{
Plot pacf and acf of a time-series.
}
