% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_plotting_func.R
\name{plot_unit_root_figs}
\alias{plot_unit_root_figs}
\title{Plot ACF and PACF for unit-roots testing.}
\usage{
plot_unit_root_figs(da_ts, freq, start, lag_max, main, xlab, ylab)
}
\arguments{
\item{da_ts}{An original ts object.}

\item{freq}{The frequency of the ts object.}

\item{start}{The start of the diff(da_ts) object.}

\item{lag_max}{The lag.max param for acf().}

\item{main}{The title of the figure.}

\item{xlab}{xlab.}

\item{ylab}{ylab.}
}
\description{
Plot ACF and PACF for unit-roots testing.
}
